Moved uploading logic to config, cleaned up routes
