+++ routes/repository.js
@@ -1,10 +1,8 @@
 const { Router } = require('express')
-const multer = require('multer')
-const path = require('path')
-const fs = require('fs')
 const asyncHandler = require('express-async-handler')
 const { authToken } = require('../middlewares/verifyToken')
 const { checkView } = require('../middlewares/checkRepoView')
+const upload = require('../config/multer') 
 
 const router = Router()
 
@@ -16,23 +14,6 @@ const {
   uploadToRepository
 } = require('../controllers/repository')
 
-const storage = multer.diskStorage({
-  destination: (req, file, cb)=> {
-    const uname = req.params.username
-    const repo = req.params.repo
-    const dir =  `./uploads/${uname}/${repo}`
-    if (!fs.existsSync(dir)) 
-      fs.mkdirSync(dir, {recursive: true})
-    cb(null, dir)
-  },
-  filename: (req, file, cb)=> {
-    console.log(file)
-    cb(null, Date.now() + path.extname(file.originalname)) //This will be the file name
-  }
-})
-//No file filter since no restriction on uploaded file extensions
-
-const upload = multer({storage: storage})
 const numFilesLimit = 20
 
 router
